+++ R/plots.R
@@ -5,6 +5,9 @@
 #' @param type The type of plot to create. Valid options are "percent" (the default) or "count".
 #' @param fill The fill of the plot. Valid options are a character color (for one variable plots) or
 #'    a variable given in formula notation (`~var`), used to create a grouped bar plot.
+#' @param orient The orientation for the plot (either "vertical", the default, or "horizontal"). As a
+#'   shortcut, "v" and "h" may be used.
+#' @param dodge The number of rows to dodge the axis labels to should they be overlapping.
 #' @param title An override for the title of the plot. A sensible default is provided.
 #' @param na_rm Should missing values be removed? Defaults to FALSE.
 #' @param ... Extra title arguments passed on to [ggformula::gf_labs()] (which feeds to [ggplot2::ggplot()]).
@@ -16,14 +19,16 @@
 #' plot_bar(mtcars, ~cyl)
 #' plot_bar(mtcars, ~cyl, type = "count")
 #' plot_bar(mtcars, ~cyl, type = "percent", fill = "yellowgreen")
+#' plot_bar(mtcars, ~cyl, orient = "horizontal")
+#' plot_bar(dplyr::starwars, ~hair_color, dodge = 2)
 #'
 #' plot_bar(mtcars, ~cyl, fill = ~gear)
 #' plot_bar(mtcars, ~cyl, type = "count", fill = ~gear)
-plot_bar <- function(data, formula, type = c("percent", "count"), fill = '#0032A0', title = NULL, na_rm = FALSE, ...) {
+plot_bar <- function(data, formula, type = c("percent", "count"), fill = '#0032A0', orient = c("vertical", "horizontal"), dodge = 1, title = NULL, na_rm = FALSE, ...) {
 
   # check for empty strings and make them actual NAs
   data <- tibble::as_tibble(data) %>%
-    dplyr::mutate(dplyr::across(where(is.character), ~dplyr::na_if(., "")))
+    dplyr::mutate(dplyr::across(dplyr::where(is.character), ~dplyr::na_if(., "")))
 
   # error catching
 
@@ -42,6 +47,7 @@ plot_bar <- function(data, formula, type = c("percent", "count"), fill = '#0032A
 
   # code
   type <- base::match.arg(type)
+  orient <- base::match.arg(orient)
   var <- formula[[2]]
   var_str <- base::deparse(base::substitute(var))
 
@@ -74,7 +80,7 @@ plot_bar <- function(data, formula, type = c("percent", "count"), fill = '#0032A
 
       }
 
-      data %>%
+      plot <- data %>%
         dplyr::mutate("{var}" := base::factor({{ var }})) %>%
         ggformula::gf_percents(formula, fill = fill) %>%
         ggformula::gf_refine(ggplot2::scale_y_continuous(expand = ggplot2::expansion(mult = c(0, 0.1)))) %>%
@@ -84,7 +90,12 @@ plot_bar <- function(data, formula, type = c("percent", "count"), fill = '#0032A
                            subtitle = base::paste("Missing:", na, "|", "NAs Removed:",
                                                   base::ifelse(na_rm == FALSE, "No", "Yes")),
                            y = "Percent", ...) %>%
+        ggformula::gf_refine(ggplot2::scale_x_discrete(guide = ggplot2::guide_axis(n.dodge = dodge))) %>%
         finalize_plot()
+
+      if (orient == "vertical") return(plot) else return(plot + ggplot2::coord_flip())
+
+
     } else {
       # grouped plot (variable fill)
 
@@ -102,7 +113,7 @@ plot_bar <- function(data, formula, type = c("percent", "count"), fill = '#0032A
 
       }
 
-      data %>%
+      plot <- data %>%
         dplyr::mutate("{var}" := base::gsub(" ", "\n", {{ var }}),
                       "{fill_var}" := base::factor({{ fill_var }})) %>%
         ggformula::gf_percents(formula,
@@ -119,8 +130,11 @@ plot_bar <- function(data, formula, type = c("percent", "count"), fill = '#0032A
                                                   "NAs Removed:", base::ifelse(na_rm == FALSE, "No", "Yes")),
                            ...) %>%
         ggformula::gf_refine(ggplot2::scale_fill_brewer(palette = "Dark2", na.value = "grey"),
-                             ggplot2::scale_y_continuous(expand = ggplot2::expansion(mult = c(0, 0.1)))) %>%
+                             ggplot2::scale_y_continuous(expand = ggplot2::expansion(mult = c(0, 0.1))),
+                             ggplot2::scale_x_discrete(guide = ggplot2::guide_axis(n.dodge = dodge))) %>%
         finalize_plot()
+
+      if (orient == "vertical") return(plot) else return(plot + ggplot2::coord_flip())
     }
 
 
@@ -140,18 +154,21 @@ plot_bar <- function(data, formula, type = c("percent", "count"), fill = '#0032A
 
       }
 
-      data %>%
+      plot <- data %>%
         dplyr::mutate("{var}" := base::factor({{ var }})) %>%
         ggformula::gf_counts(formula, fill = fill) %>%
-        ggformula::gf_refine(ggplot2::scale_y_continuous(expand = ggplot2::expansion(mult = c(0, 0.1)),
-                                                         )) %>%
+        ggformula::gf_refine(ggplot2::scale_y_continuous(expand = ggplot2::expansion(mult = c(0, 0.1)))) %>%
         ggformula::gf_labs(title = ifelse(base::is.null(title),
                                           paste0("Bar Chart (Counts) of ",
                                                  var_str), title),
                            subtitle = base::paste("Missing:", na, "|", "NAs Removed:",
                                                   base::ifelse(na_rm == FALSE, "No", "Yes")),
                            y = "Count", ...) %>%
+        ggformula::gf_refine(ggplot2::scale_x_discrete(guide = ggplot2::guide_axis(n.dodge = dodge))) %>%
         finalize_plot()
+
+      if (orient == "vertical") return(plot) else return(plot + ggplot2::coord_flip())
+
     } else {
       # grouped count (variable fill)
 
@@ -169,7 +186,7 @@ plot_bar <- function(data, formula, type = c("percent", "count"), fill = '#0032A
 
       }
 
-      data %>%
+      plot <- data %>%
         dplyr::mutate("{var}" := base::gsub(" ", "\n", {{ var }}),
                       "{fill_var}" := base::factor({{ fill_var }})) %>%
         ggformula::gf_counts(formula,
@@ -185,8 +202,11 @@ plot_bar <- function(data, formula, type = c("percent", "count"), fill = '#0032A
                                  "NAs Removed:", base::ifelse(na_rm == FALSE, "No", "Yes")),
           ...) %>%
         ggformula::gf_refine(ggplot2::scale_fill_brewer(palette = "Dark2", na.value = "grey"),
-                             ggplot2::scale_y_continuous(expand = ggplot2::expansion(mult = c(0, 0.1)))) %>%
+                             ggplot2::scale_y_continuous(expand = ggplot2::expansion(mult = c(0, 0.1))),
+                             ggplot2::scale_x_discrete(guide = ggplot2::guide_axis(n.dodge = dodge))) %>%
         finalize_plot()
+
+      if (orient == "vertical") return(plot) else return(plot + ggplot2::coord_flip())
     }
   }
 }
@@ -207,11 +227,11 @@ plot_bar <- function(data, formula, type = c("percent", "count"), fill = '#0032A
 #' plot_box(mtcars, ~wt)
 #' plot_box(mtcars, wt~gear, fill = 'orangered4')
 #' plot_box(mtcars, wt~gear)
-plot_box <- function(data, formula, fill = "grey80", title = NULL, ...) {
+plot_box <- function(data, formula, fill = "grey80", orient = c("vertical", "horizontal"), title = NULL, ...) {
 
   # check for empty strings and make them actual NAs
   data <- tibble::as_tibble(data) %>%
-    dplyr::mutate(dplyr::across(where(is.character), ~dplyr::na_if(., "")))
+    dplyr::mutate(dplyr::across(dplyr::where(is.character), ~dplyr::na_if(., "")))
 
   # error catching
   check_test(ggformula::gf_boxplot(formula, data = data, geom = "errorbar", linewidth = 2, width = 0))
@@ -219,6 +239,7 @@ plot_box <- function(data, formula, fill = "grey80", title = NULL, ...) {
   rlang::inform("Note: NAs always removed for boxplots", .frequency = "once", .frequency_id = "box-nas")
 
   # code
+  orient <- base::match.arg(orient)
   var <- formula[[2]]
   var_str <- base::deparse(base::substitute(var))
 
@@ -243,7 +264,7 @@ plot_box <- function(data, formula, fill = "grey80", title = NULL, ...) {
       stats::na.omit()
 
 
-    ggformula::gf_boxplot(formula, data = data, geom = "errorbar", linewidth = 1.5, width = 0) %>%
+    plot <- ggformula::gf_boxplot(formula, data = data, geom = "errorbar", linewidth = 1.5, width = 0) %>%
       ggformula::gf_boxplot(formula, data = data, fill = fill, width = 0.5, linewidth = 1, color = "black",
                             outlier.shape = 21, outlier.size = 2.5, outlier.color = "grey70",
                             outlier.fill = "black") %>%
@@ -261,6 +282,8 @@ plot_box <- function(data, formula, fill = "grey80", title = NULL, ...) {
                axis.text.y = ggplot2::element_blank()) +
       ggplot2::ylim(-0.4, 0.4)
 
+    if (orient == "vertical") return(plot) else return(plot + ggplot2::coord_flip())
+
 
   } else { # grouped boxplot
 
@@ -274,7 +297,7 @@ plot_box <- function(data, formula, fill = "grey80", title = NULL, ...) {
       stats::na.omit() %>%
       dplyr::mutate("{by_var}" := base::factor({{ by_var }}))
 
-    ggformula::gf_boxplot(formula, data = data, geom = "errorbar", linewidth = 2, width = 0) %>%
+    plot <- ggformula::gf_boxplot(formula, data = data, geom = "errorbar", linewidth = 2, width = 0) %>%
       ggformula::gf_boxplot(formula, data = data, fill = fill, width = 0.5, lwd = 1, color = "black",
                             outlier.shape = 21, outlier.size = 2.5, outlier.color = "grey70",
                             outlier.fill = "black", notchwidth = 2) %>%
@@ -287,6 +310,8 @@ plot_box <- function(data, formula, fill = "grey80", title = NULL, ...) {
                          ...) %>%
       finalize_plot() %>%
       ggformula::gf_refine(ggplot2::scale_y_continuous(labels = plot_labels))
+
+    if (orient == "vertical") return(plot) else return(plot + ggplot2::coord_flip())
 
   }
 
@@ -327,7 +352,7 @@ plot_hist <- function(data, formula, fill = "#0032A0", breaks = NULL, group = NU
 
   # check for empty strings and make them actual NAs
   data <- tibble::as_tibble(data) %>%
-    dplyr::mutate(dplyr::across(where(is.character), ~dplyr::na_if(., "")))
+    dplyr::mutate(dplyr::across(dplyr::where(is.character), ~dplyr::na_if(., "")))
 
   # error catching
   if (is.null(breaks)) {
@@ -385,7 +410,7 @@ plot_hist <- function(data, formula, fill = "#0032A0", breaks = NULL, group = NU
                            subtitle = base::paste("Missing:", n_na, "|", "NAs Removed: Yes"),
                            ...) %>%
         finalize_plot() %>%
-        ggformula::gf_refine(ggplot2::scale_x_continuous(labels = plot_labels, guide = guide_axis(check.overlap = TRUE)))
+        ggformula::gf_refine(ggplot2::scale_x_continuous(labels = plot_labels, guide = ggplot2::guide_axis(check.overlap = TRUE)))
 
       return(plot)
 
@@ -400,7 +425,7 @@ plot_hist <- function(data, formula, fill = "#0032A0", breaks = NULL, group = NU
                            subtitle = base::paste("Missing:", n_na, "|", "NAs Removed: Yes"),
                            ...) %>%
         finalize_plot() %>%
-        ggformula::gf_refine(ggplot2::scale_x_continuous(labels = plot_labels, breaks = breaks, guide = guide_axis(check.overlap = TRUE)))
+        ggformula::gf_refine(ggplot2::scale_x_continuous(labels = plot_labels, breaks = breaks, guide = ggplot2::guide_axis(check.overlap = TRUE)))
 
       return(plot)
 
@@ -436,7 +461,7 @@ plot_hist <- function(data, formula, fill = "#0032A0", breaks = NULL, group = NU
                                                   "NAs Removed: Yes"),
                            ...) %>%
         finalize_plot() %>%
-        ggformula::gf_refine(ggplot2::scale_x_continuous(labels = plot_labels, guide = guide_axis(angle = x_angle))) %>%
+        ggformula::gf_refine(ggplot2::scale_x_continuous(labels = plot_labels, guide = ggplot2::guide_axis(check.overlap = TRUE))) %>%
         ggformula::gf_theme(panel.border = ggplot2::element_rect(color = "black", fill = NA))
 
       return(plot)
@@ -457,7 +482,7 @@ plot_hist <- function(data, formula, fill = "#0032A0", breaks = NULL, group = NU
         finalize_plot() %>%
         ggformula::gf_refine(ggplot2::scale_x_continuous(labels = plot_labels,
                                                          breaks = breaks,
-                                                         guide = guide_axis(angle = x_angle))) %>%
+                                                         guide = ggplot2::guide_axis(check.overlap = TRUE))) %>%
         ggformula::gf_theme(panel.border = ggplot2::element_rect(color = "black", fill = NA))
 
       return(plot)
@@ -501,7 +526,7 @@ plot_scatter <- function(data, formula, fill = "#0032a0", title = NULL, legend_t
 
   # check for empty strings and make them actual NAs
   data <- tibble::as_tibble(data) %>%
-    dplyr::mutate(dplyr::across(where(is.character), ~dplyr::na_if(., "")))
+    dplyr::mutate(dplyr::across(dplyr::where(is.character), ~dplyr::na_if(., "")))
 
   # error catching
   rlang::inform("Note: NAs always removed (in pairs) for scatterplots.", .frequency = "regularly", .frequency_id = "scatter_nas")