+++ tests/testthat/test-infer_other.R
@@ -28,19 +28,22 @@ test_that("regression fails with mispelling", {
 # test_chisq() ####
 
 test_that("chisq works", {
-  expect_no_condition(
+  # skip("once per session")
+  expect_no_warning(
     suppressWarnings(infer_chisq(mtcars, cyl~gear))
   )
 })
 
 test_that("chisq works with specifying test", {
-  expect_no_condition(
+  # skip("once per session")
+  expect_no_warning(
     suppressWarnings(infer_chisq(mtcars, cyl~gear, type = "test"))
   )
 })
 
 test_that("chisq works with specifying expected", {
-  expect_no_condition(
+  # skip("once per session")
+  expect_no_warning(
     suppressWarnings(infer_chisq(mtcars, cyl~gear, type = "expected"))
   )
 })